Add files via upload
This is the Postman collection with documentation for the Formula One API hosted at Ergast API. This web service which provides a historical record of Formula one motor racing data for non-commercial purposes. Please read the terms and conditions of use. The API provides data for the Formula One series, from the beginning of the world championships in 1950.